added todos in datasets and corrected imports
